morning changes

+++ src/pages/RecipeDetails.jsx
@@ -4,6 +4,7 @@ import styles from '../styles/RecipeDetails.module.css';
 import parse from 'html-react-parser';
 import buttonStyles from '../styles/Buttons.module.css';
 import { RecipesContext } from "../components/contexts/RecipesContext";
+import { Badge} from '@mantine/core';
 
 const RecipeDetails = () => {
     const {recipe,fetchOneRecipe } = useContext(RecipesContext);
@@ -18,6 +19,9 @@ const RecipeDetails = () => {
     console.log(recipe)
     
     const handleDelete = async () => {
+      const confirmDelete = window.confirm(`Are you sure you want to delete this recipe? There's no going back once it's done`);
+
+      if (confirmDelete) {
         try {
           const response = await fetch(`${import.meta.env.VITE_BASE_API_URL}/recipes/${recipeId}`, {
             method: 'DELETE',
@@ -29,7 +33,9 @@ const RecipeDetails = () => {
         } catch (error) {
           console.error(error)
         }
+      } else { 
       }
+    }
 
   return (
     <>
@@ -39,19 +45,31 @@ const RecipeDetails = () => {
           <img src={recipe.image} alt={recipe.title} />
         </div>
         <div className={styles.HeaderboxRight}>
-          
+          <p>Ready in: {recipe.readyInMinutes} min</p>
+          <p>{recipe.servings} serving(s)</p>
+          <div>{recipe.dishTypes}</div>
           <div>
-          <h2 className={styles.h2}>Ingredients</h2>
-            <ul className={styles.ul}>
-              {recipe.extendedIngredients ?
-              recipe.extendedIngredients.map(ingredient => {return (
-              <li key={ingredient.id}>{ingredient.original}</li>)}) :<p>loading</p> }
-            </ul>
-          </div>
+            {recipe.diets && recipe.diets.length > 0 && (
+            recipe.diets.map((diet, index) => {
+            return (
+              <Badge key={index} color="#f4612d" style={{ marginRight: '4px' }}>
+              {diet}
+              </Badge>
+            );
+            })
+            )}
+           </div>
         </div>
       </div>
       <div className={styles.Mainbox}>
-      
+      <div>
+        <h2 className={styles.h2}>Ingredients</h2>
+          <ul className={styles.ul}>
+            {recipe.extendedIngredients ?
+            recipe.extendedIngredients.map(ingredient => {return (
+            <li key={ingredient.id}>{ingredient.original}</li>)}) :<p>loading</p> }
+          </ul>
+      </div>
         <h2 className={styles.h2}>Instructions</h2>
         <div>
         {console.log(recipe.instructions)}
